rename to mantain BC compatibility
